Commit before move/insert rework so it doesn't write inbetween grapheme bytes.

+++ src/charlie-utils/conversion.odin
@@ -7,7 +7,7 @@ fVec :: proc(i_vector: [$N]int) -> (f_vector: [N]f32) {
 	return f_vector
 }
 
-iVec :: proc(f_vector: [$N]int) -> (i_vector: [N]f32) {
+iVec :: proc(f_vector: [$N]f32) -> (i_vector: [N]int) {
 	for val, i in f_vector {
 		i_vector[i] = int(f_vector[i])
 	}
